# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MikroTik Centralized Fleet Management API
  version: 1.20.0
...
servers:
- url: /api/v1
paths:
...
  /services:
    get:
      summary: List ten cac logical service dang co endpoint
      tags:
      - ServiceEndpoints
      parameters:
      - &id014
        in: header
        name: x-actor-permissions
        required: false
        description: TAM THOI (AuthStubGuard, chua phai Auth/RBAC that) -- CSV permission actor duoc cap. Thieu quyen bat buoc -> 403 FORBIDDEN.
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': &id015
          description: FORBIDDEN -- thieu permission
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
  /services/{serviceName}/endpoints:
    post:
      summary: Tao endpoint moi cho mot service (KHONG rollout -- chi tao record + dang ky vao registry)
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: serviceName
        required: true
        schema:
          type: string
        description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceEndpointCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
              example: &id016
                data:
                  id: 5f83a56b-7e57-43b0-a128-856556a42528
                  service_name: radius-primary
                  service_type: RADIUS
                  environment: development
                  host: 10.20.0.11
                  port: 1812
                  protocol: UDP
                  priority: 10
                  enabled: true
                  region: null
                  ship_scope:
                    type: ALL
                  healthcheck_type: RADIUS_ACCESS_REQUEST
                  healthcheck_interval_s: 30
                  timeout_ms: 2000
                  secret_configured: true
                  check_config:
                    probe_username: probe
                    probe_password: probe
                  in_maintenance: false
                  config_version: 1
                  status: HEALTHY
                  is_active: true
                  last_check_at: '2026-08-25T01:17:25.388Z'
                  last_success_at: '2026-08-25T01:17:25.388Z'
                  last_rtt_ms: 24
                  failure_count: 0
                  breaker_state: CLOSED
                  last_error: null
                  created_at: '2026-08-25T01:16:09.140Z'
                  updated_at: '2026-08-25T01:16:09.140Z'
                meta:
                  request_id: req_01M0V7TC1GRVQTZTRHX6AK3HX9
                  generated_at: '2026-08-25T01:16:09.148Z'
                error: null
        '400': &id017
          description: VALIDATION_FAILED
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
        '403': *id015
  /service-endpoints:
    get:
      summary: List service endpoints (loc service_name/environment/enabled)
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: query
        name: service_name
        schema:
          type: string
      - in: query
        name: environment
        schema:
          type: string
      - in: query
        name: enabled
        schema:
          type: string
          enum:
          - 'true'
          - 'false'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id015
  /service-endpoints/{id}:
    get:
      summary: Get service endpoint detail (secret_configured, khong bao gio secret that)
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
              example: *id016
        '403': *id015
        '404': &id018
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
    patch:
      summary: Cap nhat endpoint -- hot-reload ServiceRegistry ngay (khong can restart), sinh revision moi
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceEndpointUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id017
        '403': *id015
        '404': *id018
    delete:
      summary: Xoa endpoint (soft delete). Chan neu la endpoint HEALTHY cuoi cung cua service.
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '204':
          description: Deleted
        '403': *id015
        '404': *id018
        '409':
          description: LAST_HEALTHY_ENDPOINT -- day la endpoint HEALTHY cuoi cung, khong the tu cat chan minh
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
              example: &id019
                data: null
                meta:
                  request_id: req_01M0V7YX51K7XE3XJNR81BGN9Y
                error:
                  code: LAST_HEALTHY_ENDPOINT
                  message: Cannot disable the last healthy endpoint of service "control-db-primary"
                  details:
                    service_name: control-db-primary
                  retryable: false
  /service-endpoints/{id}/check:
    post:
      summary: Chay synthetic health check ngay lap tuc (khong doi den chu ky sweep)
      description: Ap dung retry (exponential backoff + jitter) + timeout tren tung attempt + circuit breaker (CLOSED/OPEN/HALF_OPEN) truoc khi cap nhat trang thai. Neu ket qua la UNHEALTHY se ban/cap nhat alert tuong ung (Sec Alerts) va ghi audit_log.
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '403': *id015
        '404': *id018
        '201':
          description: Ket qua check (real functional check, khong chi TCP port)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
              example:
                data:
                  id: 5f83a56b-7e57-43b0-a128-856556a42528
                  service_name: radius-primary
                  service_type: RADIUS
                  environment: development
                  host: 10.20.0.11
                  port: 1812
                  protocol: UDP
                  priority: 10
                  enabled: true
                  region: null
                  ship_scope:
                    type: ALL
                  healthcheck_type: RADIUS_ACCESS_REQUEST
                  healthcheck_interval_s: 30
                  timeout_ms: 2000
                  secret_configured: true
                  check_config:
                    probe_username: probe
                    probe_password: probe
                  in_maintenance: false
                  config_version: 1
                  status: UNHEALTHY
                  is_active: false
                  last_check_at: '2026-08-25T01:17:25.388Z'
                  last_success_at: null
                  last_rtt_ms: null
                  failure_count: 1
                  breaker_state: OPEN
                  last_error:
                    message: RADIUS request timed out after 2000ms
                    attempts:
                    - attempt: 1
                      error: RADIUS request timed out after 2000ms
                      delayMs: 195
                    - attempt: 2
                      error: RADIUS request timed out after 2000ms
                      delayMs: 293
                    - attempt: 3
                      error: RADIUS request timed out after 2000ms
                      delayMs: 0
                  created_at: '2026-08-25T01:16:09.140Z'
                  updated_at: '2026-08-25T01:17:25.388Z'
                meta:
                  request_id: req_01M0V7WG42D2EBNV36MNAE3WK3
                  generated_at: '2026-08-25T01:17:25.407Z'
                error: null
  /service-endpoints/{id}/enable:
    post:
      summary: Bat endpoint (enabled=true)
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '403': *id015
        '404': *id018
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
  /service-endpoints/{id}/disable:
    post:
      summary: Tat endpoint (enabled=false). Chan neu la endpoint HEALTHY cuoi cung.
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '403': *id015
        '404': *id018
        '409':
          description: LAST_HEALTHY_ENDPOINT
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
              example: *id019
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
  /service-endpoints/{id}/maintenance:
    post:
      summary: Bat/tat maintenance mode (status -> MAINTENANCE, loai khoi active selection)
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id017
        '403': *id015
        '404': *id018
  /service-endpoints/{id}/revisions:
    get:
      summary: Lich su thay doi cua mot endpoint (moi PATCH la 1 revision)
      tags:
      - ServiceEndpoints
      parameters:
      - *id014
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceEndpointRevision'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id015
        '404': *id018
...
components:
...
  schemas:
    BaseMeta:
      type: object
      properties:
        request_id:
          type: string
        generated_at:
          type: string
          format: date-time
        data_from:
          type: string
          format: date-time
        data_to:
          type: string
          format: date-time
        data_freshness_seconds:
          type: integer
          nullable: true
        freshness_by_source:
          type: object
          additionalProperties:
            type: integer
            nullable: true
        page:
          type: object
          nullable: true
          description: Chi xuat hien tren cac endpoint list co phan trang (cursor hoac offset).
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
            cursor_next:
              type: string
              nullable: true
            has_more:
              type: boolean
        warnings:
          type: array
          nullable: true
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              details:
                type: object
                additionalProperties: true
...
    ErrorDetails:
      type: object
      description: Body loi chuan hoa -- AGENT_COLLABORATION.md Sec.5. Khong bao gio chua stack trace hay loi driver that.
      properties:
        code:
          type: string
          enum:
          - VALIDATION_FAILED
          - INVALID_TIME_RANGE
          - INSUFFICIENT_DATA
          - RECONCILIATION_UNAVAILABLE
          - UNAUTHENTICATED
          - FORBIDDEN
          - SCOPE_FORBIDDEN
          - AREA_NOT_FOUND
          - SHIP_NOT_FOUND
          - DEVICE_NOT_FOUND
          - INTERFACE_NOT_FOUND
          - ZONE_NOT_FOUND
          - ENDPOINT_NOT_FOUND
          - ALERT_NOT_FOUND
          - SERVICE_NOT_FOUND
          - RESOURCE_CONFLICT
          - AREA_HAS_SHIPS
          - SHIP_HAS_DEVICES
          - INTERFACE_DOUBLE_COUNT
          - LAST_HEALTHY_ENDPOINT
          - ENDPOINT_ALREADY_EXISTS
          - DATABASE_UNAVAILABLE
          - RATE_LIMITED
          - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
        retryable:
          type: boolean
        retry_after_seconds:
          type: integer
          nullable: true
      required:
      - code
      - message
      - retryable
    HealthStatus:
      type: string
      description: '5 trang thai bat buoc: HEALTHY, DEGRADED, UNHEALTHY, UNKNOWN, MAINTENANCE.'
      enum:
      - HEALTHY
      - DEGRADED
      - UNHEALTHY
      - UNKNOWN
      - MAINTENANCE
...
    ServiceEndpoint:
      type: object
      description: Mot dia chi (host/port) cu the cua mot service logic (service_name). Day la nguon su that DUY NHAT cho dia chi RADIUS/database/collector/... -- ServiceRegistry doc tu bang nay, khong co code nao duoc phep hard-code IP (ADR-04). secret KHONG BAO GIO xuat hien trong response, chi co secret_ref (tham chieu) va secret_configured (boolean).
      properties:
        id:
          type: string
          format: uuid
        service_name:
          type: string
          description: Ten service logic, vd 'radius-primary'
        service_type:
          type: string
          enum: &id051
          - RADIUS
          - RADIUS_ACCT
          - DATABASE
          - COLLECTOR
          - STORAGE
          - USER_MANAGER
          - CONTROLLER
          - BUS
          - CACHE
          - ZEROTIER
        environment:
          type: string
        host:
          type: string
        port:
          type: integer
        protocol:
          type: string
        priority:
          type: integer
          description: So nho hon = uu tien cao hon khi chon active endpoint
        enabled:
          type: boolean
        region:
          type: string
          nullable: true
        ship_scope:
          type: object
          additionalProperties: true
        healthcheck_type:
          type: string
          enum: &id052
          - RADIUS_ACCESS_REQUEST
          - RADIUS_ACCT_PROBE
          - SQL_RW
          - CLICKHOUSE_PING
          - FLOW_RECENCY
          - POLL_RECENCY
          - S3_RW
          - HTTP_FUNCTIONAL
          - NATS_RTT
          - BACKUP_FRESHNESS
        healthcheck_interval_s:
          type: integer
        timeout_ms:
          type: integer
        secret_configured:
          type: boolean
          description: true neu secret_ref resolve duoc -- KHONG BAO GIO tra secret that
        check_config:
          type: object
          additionalProperties: true
          description: Tham so rieng cho tung loai checker (vd probe_username, database, path)
        in_maintenance:
          type: boolean
        config_version:
          type: integer
          description: Tang moi lan PATCH -- doi chieu voi revisions[]
        status:
          $ref: '#/components/schemas/HealthStatus'
        is_active:
          type: boolean
          description: Endpoint dang duoc chon phuc vu (uu tien cao nhat con HEALTHY/DEGRADED)
        last_check_at:
          type: string
          format: date-time
          nullable: true
        last_success_at:
          type: string
          format: date-time
          nullable: true
        last_rtt_ms:
          type: integer
          nullable: true
        failure_count:
          type: integer
        breaker_state:
          type: string
          enum: &id053
          - CLOSED
          - OPEN
          - HALF_OPEN
        last_error:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServiceEndpointCreateRequest:
      type: object
      required:
      - service_type
      - host
      - port
      - protocol
      - healthcheck_type
      properties:
        service_type:
          type: string
          enum: *id051
        environment:
          type: string
          default: development
        host:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        protocol:
          type: string
        priority:
          type: integer
          default: 100
        region:
          type: string
          nullable: true
        ship_scope:
          type: object
          additionalProperties: true
          default:
            type: ALL
        healthcheck_type:
          type: string
          enum: *id052
        healthcheck_interval_s:
          type: integer
          default: 30
        timeout_ms:
          type: integer
          default: 5000
        secret_ref:
          type: string
          nullable: true
          description: Tham chieu secret, vd 'env:RADIUS_PROD_SECRET' (dev resolver). KHONG BAO GIO gia tri secret that.
        check_config:
          type: object
          additionalProperties: true
          default: {}
        enabled:
          type: boolean
          default: true
    ServiceEndpointUpdateRequest:
      type: object
      required:
      - reason
      description: reason BAT BUOC -- moi thay doi endpoint di thang vao audit_logs (SYSTEM_SPEC Sec.9.2).
      properties:
        host:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        protocol:
          type: string
        priority:
          type: integer
        region:
          type: string
          nullable: true
        ship_scope:
          type: object
          additionalProperties: true
        healthcheck_type:
          type: string
          enum: *id052
        healthcheck_interval_s:
          type: integer
        timeout_ms:
          type: integer
        secret_ref:
          type: string
          nullable: true
        check_config:
          type: object
          additionalProperties: true
        reason:
          type: string
          minLength: 3
    MaintenanceRequest:
      type: object
      required:
      - enable
      - reason
      properties:
        enable:
          type: boolean
        reason:
          type: string
          minLength: 3
    ServiceEndpointRevision:
      type: object
      description: Lich su thay doi endpoint -- moi PATCH tao 1 revision, dung de audit/rollback thu cong.
      properties:
        version:
          type: integer
        payload:
          type: object
          additionalProperties: true
        change_reason:
          type: string
          nullable: true
        changed_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...